+++ src/content.ts
@@ -1,3 +1,0 @@
-// This file is for content script functionality
-// Currently not used in this version of the extension
-export {};