+++ src/popup.ts
@@ -1,2 +1,0 @@
-// This file is not used anymore as we've migrated to React (popup.tsx)
-export {};